+++ lib/osf-components/addon/components/sign-up-form/component.ts
@@ -109,6 +109,7 @@ export default class SignUpForm extends Component.extend({
 
     @action
     submit() {
+        this.analytics.click('button', 'Home - sign_up');
         this.submitTask.perform();
     }
 }